requirements.txt
PyYAML>=3.10
boto>=2.48.0
mock==2.0.0
